+++ src/components/FavButton/styles.js
@@ -1,4 +1,4 @@
-import styled from "styled-components"
+import styled from 'styled-components'
 
 const Button = styled.button`
   border: none;
