+++ src/components/FavoriteBar/index.js
@@ -1,0 +1,3 @@
+import FavoriteBar from './FavoriteBar'
+
+export default FavoriteBar
