+++ src/components/Header/Header.js
@@ -1,19 +1,7 @@
-import styled from 'styled-components'
+import React from 'react'
+import Container from './styles'
 import logo from '../../images/logo.svg' 
-import Menu from '../Menu/Menu'
-
-const Container = styled.header`
-  width: 100%;
-  height: 100px;
-  background-color: #FF1C1C;
-  display: flex;
-  align-items: center;
-  justify-content: space-around;
-
-  .logo {
-    width: 12%;
-  }
-`
+import Menu from '../Menu'
 
 export default function Header() {
   return (
