+++ src/components/Header/styles.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+const Container = styled.header`
+  width: 100%;
+  height: 100px;
+  background-color: #ff1c1c;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+
+  .logo {
+    width: 12%;
+  }
+`;
+
+export default Container
