+++ src/stories/components/Header/Header.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Header from '../../../components/Header/Header';
+import Header from '../../../components/Header';
 
 export default {
   title: 'Header/Header',
