+++ src/stories/components/Menu/Menu.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Menu from '../../../components/Menu/Menu'
+import Menu from '../../../components/Menu'
 
 export default {
   title: 'Header/Menu',
